# swagger_api_documentation.yaml
openapi: 3.1.0
info:
  title: NirogPath API
  version: 1.0.0
  description: |
    REST API for the **NirogPath Healthcare & OPD Prescription Hub Platform**.
    A comprehensive healthcare management system supporting Patients, Doctors, Clinic Receptionists, and Superuser Admins.
    
    ## Authentication & Security
    Protected endpoints require a HTTP Bearer JWT token passed in the `Authorization` header:
    `Authorization: Bearer <jwt_token>`
    
    Tokens are obtained by authenticating via `POST /api/auth/login` or `POST /api/auth/register`.
    
    ## System Roles & Access Control Matrix
    - **patient**: Browse doctors/hospitals, book consultation slots, pay token deposit via wallet, track live OPD queue position, view e-prescriptions, log daily medication doses, manage family profiles, and ask prescription follow-up questions.
    - **doctor**: Access doctor queue, advance consultation state (`call-next`, `complete`), broadcast running late warnings, set location & availability schedules, write e-prescriptions, reply to patient prescription queries.
    - **receptionist**: Clinic-wide aggregated metrics across doctors, hospital map uploads, walk-in token creation, and patient transfer management.
    - **admin**: Superuser operations — register doctors, create partner hospitals, manage users, and view hospital analytics.
    
    ## User Story Mapping
    | User Story ID | Role | Description | API Endpoints |
    |---|---|---|---|
...
    
...
paths:
...
  /api/admin/hospitals/{hospital_id}/analytics:
    get:
      summary: Admin Hospital Analytics
      operationId: admin_hospital_analytics_api_admin_hospitals__hospital_id__analytics_get
      parameters:
        - name: hospital_id
          in: path
          required: true
          schema:
            type: string
            title: Hospital Id
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:

        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
  /api/reception/analytics:
    get:
      summary: Reception Analytics
      operationId: reception_analytics_api_reception_analytics_get
      parameters:
        - name: range
          in: query
          required: false
          schema:
            type: string
            default: today
            title: Range
        - name: doctor_id
          in: query
          required: false
          schema:
            type: string
            default: ""
            title: Doctor Id
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:

        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /api/auth/forgot-password:
    post:
      summary: Forgot Password
      description: |
        Always return 200 — never reveal whether the email is registered.
        If found, generate a one-time reset token valid for 1 hour.
      operationId: forgot_password_api_auth_forgot_password_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordIn"
        required: true
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:

        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /api/auth/reset-password:
    post:
      summary: Reset Password
      description: |
        Validate the reset token, update the user's password, mark the token as used,
        and return a fresh JWT so the user is auto-logged-in.
      operationId: reset_password_api_auth_reset_password_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordIn"
        required: true
      responses:
        200:
          description: Successful Response
          content:
            application/json:
              schema:

        422:
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
components:
  schemas:
...
    ForgotPasswordIn:
      properties:
        email:
          type: string
          format: email
          title: Email
      type: object
      required:
        - email
      title: ForgotPasswordIn
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ResetPasswordIn:
      properties:
        token:
          type: string
          title: Token
        new_password:
          type: string
          title: New Password
      type: object
      required:
        - token
        - new_password
      title: ResetPasswordIn
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
...